pyproject.toml
[project]
name = "whisper-transcribe"
version = "0.2.0"
description = "Audio/video transcription using OpenAI Whisper with Apple Silicon MPS support. CLI + MCP server."
requires-python = ">=3.12"
dependencies = [
    "openai-whisper",
    "torch",
    "torchvision",
    "torchaudio",
    "rich",
    "mcp",
]

[project.scripts]
transcribe = "whisper_transcribe.cli:main"
transcribe-mcp = "whisper_transcribe.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/whisper_transcribe"]
